BULK: getadmiral.com Related domains +
This commit handles approximately 500 new domains, which are contributed
with 637 records by @migueldemoura at https://github.com/easylist/easylist/pull/9770.

Later migueldemoura added yet another 163 records in this comment
(1 doublet):
- https://github.com/easylist/easylist/pull/9770#issuecomment-974265456

However we did knew about 130 of them in advance.

Non the less a HUGE thanks to migueldemoura for his contribution for a
more private internet.

Closes https://mypdns.org/my-privacy-dns/matrix/-/issues/3023

## Other related links:
- https://mypdns.org/-/snippets/42 (getadmiral.com related domains)
- https://mypdns.org/-/snippets/43 (Script to commit Admiral to GIT with
  issue refs to mypdns)
- https://github.com/easylist/easylist/pull/9770 (@migueldemoura original
  commit)
- https://github.com/funilrys/PyFunceble/issues/275 (@PyFunceble bug report)

## Bulk commits
This commit made in good faith that it is compliant with the bulk commit
ruleset:
https://mypdns.org/mypdns/support/-/wikis/Contributing#bulk-commits

4914
4752