bulk tracker import
This bulk import is done in trust to of good judgement from @piperun
@D4v1dH03 @iam-py-test3.

It all started with @D4v1dH03and issue
(https://github.com/easylist/easylist/issues/10318). That lead to a full
test of active inactive records with @PyFunceble for validating the
actuality of the list and get a picture of the source credibility.

The test returned a 50/50 (IN)active records in
https://github.com/piperun/iploggerfilter/blob/master/filterlist. A
dialogue revealed @piperun outdated records "for in case of"

Related issues:
- https://github.com/easylist/easylist/issues/10318
- https://github.com/piperun/iploggerfilter/pull/51
- https://github.com/piperun/iploggerfilter/issues/49