Removed Gitlab from readme
Since Gitlab have given over the control of their network to CloudFlare and requires full paid (with your first and second born) membership of the walled garden. we can no longer in any ways support that project as it isn't supporting a free public internet.

----

Thanks to jetBrains for sponsoring IntelliJ (Ultimate Edition)
For non-commercial open source.
This helps My Privacy DNS to develop tools and maintain the blacklists.

Signed-off-by: Spirillen <pdns@protonmail.com>