Import ShadowWhisperer
The error message you're encountering indicates that the variable unique_lines is not defined when you try to use it. This is due to the indentation of the code in your script. The unique_lines variable is defined inside the read_csv function, but it is not accessible outside of that function because of the indentation.